History for Status, fixVersions, Sprints, and StoryPoints
Now the history for Status, fixVersions, Sprints, and StoryPoints shows
respective effective value at the end of the update. Previously they
were just denoting the change done by the user while now the change is
processed, and effective value as a result of the change is captured in
history data. Big win.